cli/pyproject.toml
[project]
name = "autodeploy-cli"
version = "0.1.0"
description = "AutoDeploy CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.34.2",
    "rich>=15.0.0",
    "typer>=0.25.1",
    "PyYAML>=6.0.1",
    "websockets>=14.0",
]

[project.scripts]
ad = "autodeploy_cli.main:app"

[tool.uv]
package = true

[tool.setuptools]
packages = ["autodeploy_cli"]
